Allow the app to boot optionally without creds
In some cases, the application could be used as part of a dual scheme
where the app is used as a primary mechanism but where a failover
authentication scheme exists.

In such cases, wanting the app not to go down entirely when AWS metadata
is available, but instead returning `undefined`, can be desireable.

This patch adds a configuration value that optionally allows the app not
to fail on boot when AWS metadata isn't available.

Other changes:

- Rename 'aws_erlang' in rebar.config to 'aws', the actual app name, so
  that dialyzer and other tools can properly find it.
- Change the URL input to talk with hackney to a binary() from an
  iolist() following Dialyzer recommendations.